Rename components/containers and put reclaim in its own component/container.
