+++ src/app/tabs/newAdmin/components/index.js
@@ -1,2 +1,3 @@
 export { default as AdminTabComponent } from './AdminTabComponent';
 export { default as LeftNavComponent } from './LetftNavComponent';
+export { default as ReclaimComponent } from './ReclaimComponent';
