+++ src/app/tabs/newAdmin/containers/index.js
@@ -1,2 +1,3 @@
 export { default as AdminTabContainer } from './AdminTabContainer';
 export { default as LeftNavContainer } from './LeftNavContainer';
+export { default as ReclaimContainer } from './ReclaimContainer';
