+++ src/app/tabs/newAdmin/domainInfo/components/index.js
@@ -3,4 +3,4 @@ export { default as RenewButtonComponent } from './RenewButtonComponent';
 export { default as RenewDomainComponent } from './RenewDomainComponent';
 export { default as TransferSuccessModalComponent } from './TransferSuccessModalComponent';
 export { default as UpgradeComponent } from './UpgradeComponent';
-export { default as DomainOwnerComponent } from './DomainOwnerComponent';
+export { default as SetControllerViewComponent } from './SetControllerViewComponent';
